Move flashmem.h into include folder
