+++ Sming/system/include/flashmem.h
